video: driver: Update output format with V4L2_PIX_FMT_QC10C for 10 bit
The color format on capture plane is initially defaulted to 8bit.
During a source change if the firmware detects the content to be 10-bit,
update the output format to the default 10-bit format supported by the
driver i.e. V4L2_PIX_FMT_QC10C. Client can then enumerate the supported
formats and set as required.

Change-Id: I6864820e4313de4e7078d39d74606367e3f420d0
Signed-off-by: Dikshita Agarwal <quic_dikshita@quicinc.com>