video: driver: cleanup dead code for encoder case
For the encoder case, the vpp_cycles table is 3x8, so the valid index
range for the first dimension is 0 to 2. Using index 3 will cause an
out-of-bounds issue.

Additionally, since the lcu_size for APV, AVC, and HEVC encoders does
not exceed 32, the check for lcu_size greater than 32 can be removed.

Change-Id: Ia66baccc081f071e6b4a9056d7aaf9e4f9a470df
Signed-off-by: Renjiang Han <quic_renjiang@quicinc.com>